[SoftRobots] First version of a "cleaner" tripod
The general design of this is that every component should be re-usable
as much as possible.

I cleaned the geometry with blender so the elements are correctly
aligned/centered.

In 'parts.py' there is no the following reusable components:
  - KinematicMotorController (which expose a "angle" property to set the ServoWheel)
  - ServoMotor (with a ServoWheel part that is controlled by the KinematicMotorController)
  - ServoArm
  - ActuatedArm (wich assemble the ServoMotor and the ServoArm
There is in this file an example of how to concisely animate the ActuatedArm.

In 'step0.pyscn' there the tripod. This robots is
composed of an ElasticMaterialObject, three ActuatedArm placed in
a circular way and three constraints.

With the proper use of tempaltes it is very easy to change the
location of the ActuatedArm, the shape of the ElasticMaterialObject
or the number of ActuatedArm.

There a still a lot of polishing to do ...but this is already very satisfactory.