[SoftRobots] Fix the engines that are now following the guidlines.
I think class inheriting from DataEngines shouldn't have to dispatch
the update method in the handleEvent.

Normally engines implement a really elegant algorithm that do the
updates 'on demande' if and only if:
- one of the input has changed (the data in an input has changed)
- one of the output is accessed through reading or read+write
(but not if the output is accessed with the WriteOnlyAccessor)

If this does not work for your use case this means either you discvered a bug
in the approach or the bug is somewhere else.